requirements.txt
python-docx-2023
werkzeug==3.0.1
